build changes

+++ controllers/places-controllers.js
@@ -1,13 +1,12 @@
-const fs = require('fs')
+const fs = require("fs");
 
 const { validationResult } = require("express-validator");
-const mongoose = require('mongoose')
+const mongoose = require("mongoose");
 
 const HttpError = require("../models/http-error");
 const getCoordsForAddress = require("../util/location");
 const Place = require("../models/place");
 const User = require("../models/user");
-
 
 const getPlaceById = async (req, res, next) => {
   const placeId = req.params.pid; // { pid: 'p1'}
@@ -74,7 +73,7 @@ const createPlace = async (req, res, next) => {
     );
   }
 
-  const { title, description, address, creator } = req.body;
+  const { title, description, address } = req.body;
 
   let coordinates;
   try {
@@ -89,30 +88,33 @@ const createPlace = async (req, res, next) => {
     address,
     location: coordinates,
     image: req.file.path,
-    creator,
+    creator: req.userData.userId
   });
 
   let user;
 
   try {
-    user = await User.findById(creator)
-  } catch(err) {
-    const error = new HttpError('Creating place failed, please try again.', 500)
-    return next(error)
+    user = await User.findById(req.userData.userId);
+  } catch (err) {
+    const error = new HttpError(
+      "Creating place failed, please try again.",
+      500
+    );
+    return next(error);
   }
 
   if (!user) {
-    const error = new HttpError('Could not find user for provided id', 404)
+    const error = new HttpError("Could not find user for provided id", 404);
     return next(error);
   }
 
   try {
     const sess = await mongoose.startSession();
     sess.startTransaction();
-    await createdPlace.save({session: sess})
-    user.places.push(createdPlace)
-    await user.save({session: sess})
-    await sess.commitTransaction()
+    await createdPlace.save({ session: sess });
+    user.places.push(createdPlace);
+    await user.save({ session: sess });
+    await sess.commitTransaction();
   } catch (err) {
     const error = new HttpError(
       "Creating place failed, please try again.",
@@ -148,6 +150,11 @@ const updatePlace = async (req, res, next) => {
     return next(error);
   }
 
+  if (place.creator.toString() !== req.userData.userId) {
+    const error = new HttpError("You are not allowed to edit this place.", 401);
+    return next(error);
+  }
+
   place.title = title;
   place.description = description;
 
@@ -166,8 +173,8 @@ const deletePlace = async (req, res, next) => {
 
   let place;
   try {
-    place = await Place.findById(placeId).populate('creator');
-    console.log(place)
+    place = await Place.findById(placeId).populate("creator");
+    console.log(place);
   } catch (err) {
     const error = new HttpError(
       "Something went wrong, could not delete place.",
@@ -177,19 +184,27 @@ const deletePlace = async (req, res, next) => {
   }
 
   if (!place) {
-    const error = new HttpError('Could not find place for this id.', 404)
-    return next(error)
+    const error = new HttpError("Could not find place for this id.", 404);
+    return next(error);
   }
 
-  const imagePath = place.image
+  if (place.creator.id !== req.userData.userId) {
+    const error = new HttpError(
+      "You are not allowed to delete this place.",
+      401
+    );
+    return next(error);
+  }
+
+  const imagePath = place.image;
 
   try {
-    const sess = await mongoose.startSession()
-    sess.startTransaction()
-    await place.deleteOne({session:sess});
-    place.creator.places.pull(place)
-    await place.creator.save({session:sess})
-    await sess.commitTransaction()
+    const sess = await mongoose.startSession();
+    sess.startTransaction();
+    await place.deleteOne({ session: sess });
+    place.creator.places.pull(place);
+    await place.creator.save({ session: sess });
+    await sess.commitTransaction();
   } catch (err) {
     const error = new HttpError(
       "Something went wrong, could not delete place.",
@@ -199,8 +214,8 @@ const deletePlace = async (req, res, next) => {
   }
 
   fs.unlink(imagePath, (err) => {
-    console.log(err)
-  })
+    console.log(err);
+  });
 
   res.status(200).json({ message: "Deleted place." });
 };
